src/e/edit/TagReader.java: tell Ctags what we think the file's type is, rather than trying to arrange the temporary file name so that it guesses the same as we did. This works better if a file's type changes (because you're rewriting a script from Perl to Ruby, say), avoids the possibility of disagreement, and -- as in simonj's complaint -- means that we can cope with, say, a Ruby script with no extension (where before we'd have lied to Ctags and said it was a .txt file).
src/e/edit/TagsUpdater.java: don't try to make temporary files with appropriate extensions, and don't bother calling Ctags on languages we don't think it supports (because even if it does, we won't understand its response). Try harder to ensure that the user actually sees any error messages we output, and -- until we have a better idea -- show "(No symbols available.)" for file types Ctags doesn't support.

src/e/edit/ETextWindow.java: if Evergreen's idea of the file's type changes, we need to get Ctags to take another look at it.

+++ evergreen/trunk/src/e/edit/TagsUpdater.java
@@ -244,16 +244,6 @@ public class TagsUpdater {
             return treeModel;
         }
         
-        public String getFilenameSuffix() {
-            String suffix = ".txt";
-            String filename = FileUtilities.fileFromString(getTextWindow().getFilename()).getName();
-            int lastDot = filename.lastIndexOf('.');
-            if (lastDot != -1) {
-                suffix = filename.substring(lastDot);
-            }
-            return suffix;
-        }
-        
         public void tagFound(TagReader.Tag tag) {
             publish(tag);
         }
@@ -283,26 +273,33 @@ public class TagsUpdater {
         
         public void taggingFailed(Exception ex) {
             successful = false;
-            Evergreen.getInstance().getTagsPanel().showError("<b>Is Exuberant Ctags installed and on your path?</b><p>There was an error reading the tags: " + ex.getMessage());
+            Evergreen.getInstance().getTagsPanel().showError("<b>Is Exuberant Ctags installed and on your path?</b><p>There was an error reading its output: " + ex.getMessage());
         }
         
         public void scanTags() {
             try {
+                FileType fileType = getTextWindow().getFileType();
+                if (TagReader.ctagsLanguageForFileType(fileType) == null) {
+                    Evergreen.getInstance().getTagsPanel().showError("(No symbols available.)");
+                    successful = false;
+                    return;
+                }
+                
                 // Ctags writes the name of the file into its output, so we need to use the same file each time so that md5 hashes can be compared.
                 if (temporaryFile == null) {
-                    // It's important to use the same suffix as the original file, because that's how ctags guesses the file's type, and hence which parser to use.
-                    temporaryFile = File.createTempFile("e.edit.TagsUpdater-", getFilenameSuffix());
+                    temporaryFile = File.createTempFile("e.edit.TagsUpdater-", "");
                     temporaryFile.deleteOnExit();
                 }
                 getTextArea().getTextBuffer().writeToFile(temporaryFile);
                 String charsetName = (String) getTextArea().getTextBuffer().getProperty(PTextBuffer.CHARSET_PROPERTY);
-                TagReader tagReader = new TagReader(temporaryFile, getTextWindow().getFileType(), charsetName, tagsDigest, this);
+                TagReader tagReader = new TagReader(temporaryFile, fileType, charsetName, tagsDigest, this);
                 newDigest = tagReader.getTagsDigest();
                 // See the comment above for why we don't delete our temporary files, except on exit.
                 //temporaryFile.delete();
             } catch (Exception ex) {
                 Evergreen.getInstance().getTagsPanel().showError("Couldn't make tags: " + ex.getMessage());
                 Log.warn("Couldn't make tags", ex);
+                successful = false;
             }
         }
